Update Rules.md
Reworded the rule involving excessive alts. You're not allowed to have any thus the word excessive was removed to improve the clarity of the rules, Line 11 was adjusted as well to include auto-clickers and other external applications to again improve the clarity of the rules as I have noticed at least 3 uses of auto clickers without them knowing it was against the rules, this makes it explicitly clear that these applications and tools are not allowed.