Update sam1.md
Aligned the text.
Found this entry ingame multiple times and it was all messed up, a continuous string of text that would make you screoll all the way to the right.
Tried to find the original creator of this entry to ask him to either fix it or show him what i've done, but even with Krolik i couldnt find the original creator, so i hope he doesnt mind. 
I also have a screen recording of that entry and how it looks ingame.